+++ scripts/fedora38/dockerize.sh
@@ -37,7 +37,7 @@ umount /run
 systemd-tmpfiles --create --boot
 
 # Make sure login works
-rm /var/run/nologin
+[ -f /var/run/nologin ] && rm /var/run/nologin
 
 # Mark the docker box build time.
 date --utc > /etc/docker_box_build_time
@@ -92,12 +92,12 @@ tar --create --numeric-owner --preserve-permissions --one-file-system \
 
 if [ $? != 0 ] || [ ! -f /tmp/$PACKER_BUILD_NAME.tar ]; then
   printf "\n\nTarball generation failed.\n\n"
-  printf "locked\n" | passwd --stdin root
   passwd --unlock root
+  printf "locked\nlocked\n" | passwd root
   exit 1
 else
   printf "\nTarball generation succeeded.\n"
 fi
 
-printf "locked\nlocked\n" | passwd root
 passwd --unlock root
+printf "locked\nlocked\n" | passwd root
